+++ tests/ExchangeRateProvider/ConfigurableProviderTest.php
@@ -49,13 +49,11 @@ class ConfigurableProviderTest extends AbstractTestCase
 
     private function getExchangeRateProvider(): ExchangeRateProvider
     {
-        $provider = new ConfigurableProvider();
-
-        $provider->setExchangeRate('USD', 'EUR', '0.8');
-        $provider->setExchangeRate('USD', 'GBP', '0.6');
-        $provider->setExchangeRate('USD', 'CAD', '1.2');
-        $provider->setExchangeRate('USD', 'BSD', 1);
-
-        return $provider;
+        return ConfigurableProvider::builder()
+            ->addExchangeRate('USD', 'EUR', '0.8')
+            ->addExchangeRate('USD', 'GBP', '0.6')
+            ->addExchangeRate('USD', 'CAD', '1.2')
+            ->addExchangeRate('USD', 'BSD', 1)
+            ->build();
     }
 }